requirements.txt
base58==2.1.1
cryptography==44.0.2
google-cloud-firestore==2.21.0
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
solana==0.36.6
solders==0.26.0
